Move MainPhoto in components
